chore(base): Move `Room::state` and sibling types into new `state` module.
This patch moves the `Room::state` method, along with the ``RoomState`
and `RoomStateFilter` types into the new `state` module. This patch also
moves the tests in this new module.

The idea is to group API by “theme” to get smaller modules and more
organised code.
